+++ test/ethereum/Composer.t.sol
@@ -31,7 +31,7 @@ contract ComposerTestEthereum is DeltaSetup {
         data = abi.encodePacked(transfer, data);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -63,14 +63,13 @@ contract ComposerTestEthereum is DeltaSetup {
         );
         data = abi.encodePacked(transfer, data);
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
 
         assertApproxEqAbs(amount, getCollateralBalance(user, asset, lenderId), 0);
     }
-
 
     function test_ethereum_composer_borrow(uint8 lenderId) external {
         address user = testUser;
@@ -90,7 +89,7 @@ contract ComposerTestEthereum is DeltaSetup {
 
         bytes memory data = borrow(borrowAsset, user, borrowAmount, lenderId, DEFAULT_MODE);
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -113,7 +112,7 @@ contract ComposerTestEthereum is DeltaSetup {
 
         _borrow(borrowAsset, user, borrowAmount, lenderId);
 
-        uint256 repayAmount = 20.50e6;
+        uint256 repayAmount = 20.5e6;
 
         bytes memory transfer = transferIn(
             borrowAsset,
@@ -134,7 +133,7 @@ contract ComposerTestEthereum is DeltaSetup {
         IERC20All(borrowAsset).approve(address(brokerProxyAddress), repayAmount);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -156,7 +155,7 @@ contract ComposerTestEthereum is DeltaSetup {
 
         _borrow(borrowAsset, user, borrowAmount, lenderId);
 
-        uint256 repayAmount = 120.50e6;
+        uint256 repayAmount = 120.5e6;
         deal(borrowAsset, user, repayAmount);
         bytes memory transfer = transferIn(
             borrowAsset,
@@ -176,7 +175,7 @@ contract ComposerTestEthereum is DeltaSetup {
         IERC20All(borrowAsset).approve(address(brokerProxyAddress), repayAmount);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -194,13 +193,13 @@ contract ComposerTestEthereum is DeltaSetup {
 
         _deposit(asset, user, amount, lenderId);
 
-        uint256 withdrawAmount = 2.50e18;
+        uint256 withdrawAmount = 2.5e18;
 
         bytes memory data = withdraw(asset, user, withdrawAmount, lenderId);
         approveWithdrawal(user, asset, withdrawAmount, lenderId);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -223,7 +222,7 @@ contract ComposerTestEthereum is DeltaSetup {
         approveWithdrawal(user, asset, withdrawAmount, lenderId);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -268,7 +267,7 @@ contract ComposerTestEthereum is DeltaSetup {
         IERC20All(assetIn).approve(address(brokerProxyAddress), amount);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -319,7 +318,7 @@ contract ComposerTestEthereum is DeltaSetup {
     function test_ethereum_composer_multi_route_exact_in_native() external {
         address user = testUser;
         uint256 amount = 4000.0e18;
-        uint256 amountMin = 0.10e18;
+        uint256 amountMin = 0.1e18;
 
         address assetIn = WETH;
         address assetOut = WETH;
@@ -353,7 +352,7 @@ contract ComposerTestEthereum is DeltaSetup {
 
         data = abi.encodePacked(wrap(amount), data);
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose{value: amount}(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -399,17 +398,17 @@ contract ComposerTestEthereum is DeltaSetup {
         vm.prank(user);
         IERC20All(assetIn).approve(brokerProxyAddress, amountMax);
 
-        uint balanceOutBefore = user.balance;
-        uint balanceInBefore = IERC20All(assetIn).balanceOf(user);
+        uint256 balanceOutBefore = user.balance;
+        uint256 balanceInBefore = IERC20All(assetIn).balanceOf(user);
         {
             vm.prank(user);
-            uint gas = gasleft();
+            uint256 gas = gasleft();
             IFlashAggregator(brokerProxyAddress).deltaCompose(data);
             gas = gas - gasleft();
             console.log("gas-exactOut-native-out-2 split", gas);
         }
-        uint balanceOutAfter = user.balance;
-        uint balanceInAfter = IERC20All(assetIn).balanceOf(user);
+        uint256 balanceOutAfter = user.balance;
+        uint256 balanceInAfter = IERC20All(assetIn).balanceOf(user);
 
         assertApproxEqAbs(balanceOutAfter - balanceOutBefore, amount, 1);
         assertApproxEqAbs(balanceInBefore - balanceInAfter, 665692566352012255, 0);
@@ -452,17 +451,17 @@ contract ComposerTestEthereum is DeltaSetup {
 
         data = abi.encodePacked(wrap(amountMax), data, unwrap(user, 0, ComposerUtils.SweepType.VALIDATE));
 
-        uint balanceOutBefore = IERC20All(assetOut).balanceOf(user);
-        uint balanceInBefore = user.balance;
+        uint256 balanceOutBefore = IERC20All(assetOut).balanceOf(user);
+        uint256 balanceInBefore = user.balance;
         {
             vm.prank(user);
-            uint gas = gasleft();
+            uint256 gas = gasleft();
             IFlashAggregator(brokerProxyAddress).deltaCompose{value: amountMax}(data);
             gas = gas - gasleft();
             console.log("gas-exactOut-native-in-2 split", gas);
         }
-        uint balanceOutAfter = IERC20All(assetOut).balanceOf(user);
-        uint balanceInAfter = user.balance;
+        uint256 balanceOutAfter = IERC20All(assetOut).balanceOf(user);
+        uint256 balanceInAfter = user.balance;
 
         assertApproxEqAbs(balanceOutAfter - balanceOutBefore, amount, 1);
         assertApproxEqAbs(balanceInBefore - balanceInAfter, 6303318979812611491310, 0);
@@ -508,17 +507,17 @@ contract ComposerTestEthereum is DeltaSetup {
         vm.prank(user);
         IERC20All(assetIn).approve(brokerProxyAddress, amount);
 
-        uint balanceOutBefore = user.balance;
-        uint balanceInBefore = IERC20All(assetIn).balanceOf(user);
+        uint256 balanceOutBefore = user.balance;
+        uint256 balanceInBefore = IERC20All(assetIn).balanceOf(user);
         {
             vm.prank(user);
-            uint gas = gasleft();
+            uint256 gas = gasleft();
             IFlashAggregator(brokerProxyAddress).deltaCompose(data);
             gas = gas - gasleft();
             console.log("gas-exactIn-native-out-2 split", gas);
         }
-        uint balanceOutAfter = user.balance;
-        uint balanceInAfter = IERC20All(assetIn).balanceOf(user);
+        uint256 balanceOutAfter = user.balance;
+        uint256 balanceInAfter = IERC20All(assetIn).balanceOf(user);
 
         assertApproxEqAbs(balanceOutAfter - balanceOutBefore, 11668752556768511510064, 1);
         assertApproxEqAbs(balanceInBefore - balanceInAfter, amount, 0);
@@ -567,7 +566,7 @@ contract ComposerTestEthereum is DeltaSetup {
         IERC20All(assetIn).approve(brokerProxyAddress, amount);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(abi.encodePacked(transfer, data));
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -610,7 +609,7 @@ contract ComposerTestEthereum is DeltaSetup {
         IERC20All(assetIn).approve(address(brokerProxyAddress), maxIn * 2);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -619,7 +618,7 @@ contract ComposerTestEthereum is DeltaSetup {
     function getSpotExactInMultiGen2(address[] memory tokens, uint8[] memory pids, uint16[] memory fees) internal view returns (bytes memory data) {
         uint8[] memory actions = new uint8[](pids.length);
         data = abi.encodePacked(tokens[0]);
-        for (uint i; i < pids.length; i++) {
+        for (uint256 i; i < pids.length; i++) {
             address pool = testQuoter.v3TypePool(tokens[i], tokens[i + 1], fees[i], pids[i]);
             data = abi.encodePacked(
                 data,
@@ -636,7 +635,7 @@ contract ComposerTestEthereum is DeltaSetup {
     function getSpotExactOutMultiGen2(address[] memory tokens, uint8[] memory pids, uint16[] memory fees) internal view returns (bytes memory data) {
         uint8[] memory actions = new uint8[](pids.length);
         data = abi.encodePacked(tokens[0]);
-        for (uint i; i < pids.length; i++) {
+        for (uint256 i; i < pids.length; i++) {
             actions[i] = 0;
             address pool = testQuoter.v3TypePool(tokens[i], tokens[i + 1], fees[i], pids[i]);
             data = abi.encodePacked(data, actions[i], pids[i], pool, fees[i], tokens[i + 1]);
@@ -676,7 +675,7 @@ contract ComposerTestEthereum is DeltaSetup {
         data = abi.encodePacked(transfer, data);
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
@@ -694,7 +693,7 @@ contract ComposerTestEthereum is DeltaSetup {
         );
 
         vm.prank(user);
-        uint gas = gasleft();
+        uint256 gas = gasleft();
         IFlashAggregator(brokerProxyAddress).deltaCompose(data);
         gas = gas - gasleft();
         console.log("gas", gas);
